TK_Chat/backend/requirements.txt
fastapi
uvicorn[standard]
openai==0.28
sqlalchemy
pydantic
python-dotenv
aiofiles
requests
sendgrid
apscheduler
